Specify baseline behavior in case of non-zero textangle
Currently the baseline property is underspecified: the polyline refers
to "coordinate system of the line" which is not defined anywhere else in
the document. This makes it unclear how baseline should be specified
when textangle is non-zero.

The interpretation that textangle should be ignored would result in
error in the specification because completely vertical text would have
slope angle equal to positive or negative infinity which can not be
represented by the current grammar.

Therefore, textangle should be taken into account. However, it's not
clear how it should affect baseline because the specification does not
constrain textangle to any specific angle.

This issue is fixed by explicitly specifying what coordinate system will
be used for the baseline polynomial in all possible values of textangle
property.